rename: ThumbnailBox.jsx 디렉토리 변경
